teaching_ai/otolingo/requirements.txt
firebase_functions~=0.1.0
firebase-admin~=6.2.0
fastapi~=0.104.0
pydantic~=2.4.2
python-dotenv~=1.0.0
openai~=1.3.0
asgiref~=3.7.2
functions-framework~=3.4.0
uvicorn~=0.23.2
